+++ src/main/kotlin/kscript/app/code/Templates.kt
@@ -1,7 +1,7 @@
 package kscript.app.code
 
 import kscript.app.model.PackageName
-import kscript.app.util.ScriptUtils.dropExtension
+import kscript.app.model.ScriptType
 import org.intellij.lang.annotations.Language
 
 object Templates {
@@ -46,40 +46,34 @@ object Templates {
             """.trimIndent()
     }
 
-    fun runConfig(rootNode: String, userArgs: List<String>): String {
-        val fileWithoutExtension = rootNode.dropExtension()
+    fun runConfig(rootScriptName: String, rootScriptType: ScriptType, userArgs: List<String>): String {
+        val rootFileName = rootScriptName + rootScriptType.extension
 
-        val runConfigurationBody = if (rootNode.endsWith(".kt")) {
-            """|<configuration name="$fileWithoutExtension" type="JetRunConfigurationType">
-               |    <module name="${rootNode}.main" />
-               |    <option name="VM_PARAMETERS" value="" />
-               |    <option name="PROGRAM_PARAMETERS" value="" />
-               |    <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="false" />
-               |    <option name="ALTERNATIVE_JRE_PATH" />
-               |    <option name="PASS_PARENT_ENVS" value="true" />
-               |    <option name="MAIN_CLASS_NAME" value="${fileWithoutExtension.replaceFirstChar { it.titlecase() }}Kt" />
-               |    <option name="WORKING_DIRECTORY" value="" />
-               |    <method v="2">
-               |        <option name="Make" enabled="true" />
-               |    </method>
-               |</configuration>""".trimMargin()
-        } else {
-            """|<configuration default="false" name="Main" type="BashConfigurationType" factoryName="Bash">
-               |    <option name="INTERPRETER_OPTIONS" value="" />
-               |    <option name="INTERPRETER_PATH" value="kscript" />
-               |    <option name="PROJECT_INTERPRETER" value="false" />
-               |    <option name="WORKING_DIRECTORY" value="" />
-               |    <option name="PARENT_ENVS" value="true" />
-               |    <option name="SCRIPT_NAME" value="${'$'}PROJECT_DIR${'$'}/src/${rootNode}" />
-               |    <option name="PARAMETERS" value="${userArgs.joinToString(" ")}" />
-               |    <module name="" />
-               |    <method v="2" />
-               |</configuration>""".trimMargin()
+        if (rootScriptType == ScriptType.KT) {
+            return """  |<component name="ProjectRunConfigurationManager">
+                        |  <configuration default="false" name="$rootFileName" type="JetRunConfigurationType" nameIsGenerated="true">
+                        |    <option name="MAIN_CLASS_NAME" value="${rootScriptName}Kt" />
+                        |    <module name="idea" />
+                        |    <shortenClasspath name="NONE" />
+                        |    <method v="2">
+                        |      <option name="Make" enabled="true" />
+                        |    </method>
+                        |  </configuration>
+                        |</component>
+                        |""".trimMargin()
         }
 
-        return """|<component name="ProjectRunConfigurationManager">
-                  |    $runConfigurationBody
-                  |</component>""".trimMargin()
+        return """  |<component name="ProjectRunConfigurationManager">
+                    |  <configuration default="false" name="$rootFileName" type="KotlinStandaloneScriptRunConfigurationType" nameIsGenerated="true">
+                    |    <module name="idea" />
+                    |    <shortenClasspath name="NONE" />
+                    |    <option name="filePath" value="${'$'}PROJECT_DIR${'$'}/src/$rootFileName" />
+                    |    <method v="2">
+                    |      <option name="Make" enabled="true" />
+                    |    </method>
+                    |  </configuration>
+                    |</component>
+                    |""".trimMargin()
     }
 
     fun usageOptions(selfName: String, version: String) = """