+++ src/main/kotlin/kscript/app/util/FileUtils.kt
@@ -1,6 +1,7 @@
 package kscript.app.util
 
 import kscript.app.model.OsType
+import kscript.app.model.ScriptType
 import java.io.IOException
 import java.nio.file.Files
 import java.nio.file.Path
@@ -83,5 +84,17 @@ object FileUtils {
 
         //osType == OsType.CYGWIN
         TODO()
+    }
+
+    fun resolveUniqueFilePath(basePath: Path, fileName: String, scriptType: ScriptType): Path {
+        var path = basePath.resolve(fileName + scriptType.extension)
+
+        var counter = 1
+        while (path.exists()) {
+            path = basePath.resolve(fileName + "_$counter" + scriptType.extension)
+            counter++
+        }
+
+        return path
     }
 }